Fixes consistent test failures & crashes for Unit tests.
Also fixes frequent test failure in LogTest destructor.

For PortAlarmTest, it is a simple fix to prevent consistent crash (that's now reduced to test failure).
It is more involved to fix the test failures in that file.

The README is also updated with known test failures and crashes for unit tests.